Resolved requested changes

+++ Matchismo/Model/Card/Card.m
@@ -9,7 +9,7 @@
 
 @implementation Card
 
-- (int)match:(NSArray *)otherCards {
+- (NSInteger)match:(NSArray *)otherCards {
     int score = 0;
     
     for (Card *card in otherCards) {
